Fixed double-underline by drawing lines in reverse
The second underline is drawn just below the character cell, so it
was being overwritten by the subsequent line's background (which
wasn't a problem early on when the background was just the result
of a full screen clear). Rendering lines from bottom to top means
that the second underline will now be drawn on top.