Added support for HP-style "cursor to lowerleft"
XTerm only interprets this code is hpLowerleftBugCompat is enabled
for compatibility with HP's xdb, which XTerm's documentation says
ignores termcap and sends this sequence no matter what.

I'm not sure why XTerm requires this to be explicitly enabled, as
there seems to be no other meaning to the ESC F sequence. As such,
this is being implemented as always enabled.